Fix premature capability work loop cancellation
The problem was with how the host garbage-collected expired timers (m_networkTimers) in Host::run - it checked a timer's expiration timer to see if it was expired and deleted "expired" timers, but there's a race condition where a timer's expiration time has been met and the handler is in the work queue but hasn't been executed yet, in which case the timer is deleted and the handler is run with error code operation_aborted.

To fix this, I create a dedicated steady timer (I chose steady timer over deadline timer due to its compatibility with std::chrono) for each capability in the host's m_capabilities map when registering a new capability and use that timer to schedule the capability's work loop. I also replaced the host::scheduleExecution function with host::scheduleCapabilityBackgroundWork which is used by the capabilities to schedule their background work loop and added host::postCapabilityWork so capabilities can execute work on the network thread. I also cancel timers on capability shutdown.